[refactor] Comment 리팩토링

+++ server/src/main/java/com/moovda_project/moovda/module/comment/dto/CommentPatchDto.java
@@ -10,5 +10,5 @@ public class CommentPatchDto {
     private long commentId;
 
     private String content;
-    private float star;
+    private double star;
 }
